Refactor: FirestoreLetterManager 내 storage 코드 제거 및 이관
