+++ Kabinett/Service/Letter/DefaultPhotoLetterUseCase.swift
@@ -10,16 +10,25 @@ import os
 
 final class DefaultPhotoLetterUseCase {
     private let logger: Logger
+    private let authManager: AuthManager
+    private let writerManager: FirestoreWriterManager
     private let letterManager: FirestoreLetterManager
+    private let letterStorageManager: FirestoreLetterManager
     
     init(
-        letterManager: FirestoreLetterManager
+        authManager: AuthManager,
+        writerManager: FirestoreWriterManager,
+        letterManager: FirestoreLetterManager,
+        letterStorageManager: FirestoreLetterManager
     ) {
         self.logger = Logger(
             subsystem: "co.kr.codegrove.Kabinett",
             category: "DefaultPhotoLetterUseCase"
         )
+        self.authManager = authManager
+        self.writerManager = writerManager
         self.letterManager = letterManager
+        self.letterStorageManager = letterStorageManager
     }
 }
 
@@ -38,7 +47,7 @@ extension DefaultPhotoLetterUseCase: ComponentsUseCase {
                     isRead: Bool
     ) async -> Result<Bool, any Error> {
         do {
-            let photoContentStringUrl = try await letterManager.convertPhotoToUrl(photoContents: photoContents)
+            let photoContentStringUrl = try await letterStorageManager.convertPhotoToUrl(photoContents: photoContents)
             
             let letter = Letter(
                 id: nil,
